Segundo main borrado, Metodo modificar ruta separado, Nombres cambiados
