+++ Biblioteca/src/main/java/reto/VisualizarHtml.java
@@ -1,6 +1,5 @@
 package reto;
 
-import java.awt.EventQueue;
 import java.io.File;
 
 import javax.swing.JEditorPane;
@@ -15,20 +14,8 @@ public class VisualizarHtml extends JFrame {
 	private JEditorPane editorPane;
 	Errores log = new Errores();
 
-	public static void main(String[] args) {
-		EventQueue.invokeLater(new Runnable() {
-			public void run() {
-				try {
-					VisualizarHtml frame = new VisualizarHtml();
-					frame.setVisible(true);
-				} catch (Exception e) {
-					e.printStackTrace();
-				}
-			}
-		});
-	}
-
 	public VisualizarHtml() {
+
 		setResizable(false);
 		setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
 		setBounds(100, 100, 658, 419);
